+++ src/components/ContactInputs/ContactInput.css
@@ -1,0 +1,7 @@
+.ContactInput {
+    height: 100vh;
+}
+.ContactInput p {
+font-size: 600%;
+font-weight: 400;
+}
