+++ src/components/Navbar/NavbarLogin.css
@@ -1,6 +1,7 @@
 .navbar {
-    width: 100%;
+    width: 50%;
     position: fixed;
+    background-color: rgba(255, 255, 255, 0);
 }
 
 .link-navbar {
